Tencent streaming application Seagull's logs
These are logs of Tencent streaming application Seagull. Seagull is a
streaming application which is used to collect and report real-time
click traffic of Tencent social advertisements. Seagull is architectured
on Spark Streaming. And Seagull run in a multi-tenant environment.
YARN(Hadoop-2.2.0) is the resource management platform. Seagull requests
64 executor instances with total 320 cores (each executor has 5 cores
and 12GB memory). The CPU type is Intel Xeon CPU E5-2420. The Spark
version is spark-2.1.1. And stream data source is from HDFS.